+++ packages/jest-snapshot/src/utils.ts
@@ -10,7 +10,10 @@ import chalk = require('chalk');
 import * as fs from 'graceful-fs';
 import naturalCompare = require('natural-compare');
 import type {Config} from '@jest/types';
-import {format as prettyFormat} from 'pretty-format';
+import {
+  OptionsReceived as PrettyFormatOptions,
+  format as prettyFormat,
+} from 'pretty-format';
 import {getSerializers} from './plugins';
 import type {SnapshotData} from './types';
 
@@ -152,13 +155,18 @@ export const removeLinesBeforeExternalMatcherTrap = (stack: string): string => {
 const escapeRegex = true;
 const printFunctionName = false;
 
-export const serialize = (val: unknown, indent = 2): string =>
+export const serialize = (
+  val: unknown,
+  indent = 2,
+  formatOverrides: PrettyFormatOptions = {},
+): string =>
   normalizeNewlines(
     prettyFormat(val, {
       escapeRegex,
       indent,
       plugins: getSerializers(),
       printFunctionName,
+      ...formatOverrides,
     }),
   );